Reorganized unused Autonomous classes Migrated hardware device identifiers to use strings defined in DRIFTConstants Prepared autonomous code for submersible parking code Added framework for ArmElbow RoadRunner action.
