+++ Team6220/src/main/java/org/firstinspires/ftc/team6220/old/ParkAutoLeft.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.team6220;
+package org.firstinspires.ftc.team6220.old;
 
 import com.acmerobotics.dashboard.canvas.Canvas;
 import com.acmerobotics.dashboard.telemetry.TelemetryPacket;
@@ -8,37 +8,30 @@ import com.acmerobotics.roadrunner.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
-import org.firstinspires.ftc.team6220.javatextmenu.MenuInput;
-import org.firstinspires.ftc.team6220.javatextmenu.TextMenu;
+import org.firstinspires.ftc.team6220.BaseOpMode;
+import org.firstinspires.ftc.team6220.DRIFTConstants;
 import org.firstinspires.ftc.team6220.roadrunner.MecanumDrive;
 
 /**
  * This class exposes the competition version of Autonomous. As a general rule, add code to the
  * BaseOpMode class rather than here so that it can be shared between both TeleOp and Autonomous.
  */
-
 @Disabled
-@Autonomous(name="ScoringAutoMiddle", group="Competition", preselectTeleOp="CompetitionTeleOp")
-public class ScoringAutoMiddle extends BaseOpMode {
-
+@Autonomous(name="ParkAutoLeft", group="Competition", preselectTeleOp="CompetitionTeleOp")
+public class ParkAutoLeft extends BaseOpMode {
 
     @Override
     public void runOpMode() {
 
-        Pose2d middlePose = DRIFTConstants.MIDDLE_STARTING_POSE;
+        Pose2d leftPose = DRIFTConstants.LEFT_STARTING_POSE;
 
+        MecanumDrive drive = new MecanumDrive(hardwareMap, telemetry, gamepad1, leftPose);
 
-        MecanumDrive drive = new MecanumDrive(hardwareMap, telemetry, gamepad1, middlePose);
-
-        Action middleScoringTrajectory     = drive.actionBuilder(middlePose)
-                .splineTo(new Vector2d(48, 36), (3*Math.PI)/2)
-                .endTrajectory()
-                .splineTo(new Vector2d(48, 50), (5*Math.PI)/4) //scored a sample
-                .splineTo(new Vector2d(25, 10),(3*Math.PI)/2)
-                .splineTo(new Vector2d(28, 10),(1*Math.PI)/1)
+        // Create left parking trajectory
+        Action leftParkingTrajectory = drive.actionBuilder(leftPose)
+                .splineTo(new Vector2d(-60, 60), (Math.PI))
                 .build();
-
-        Action trajectoryAction = middleScoringTrajectory;
+        Action trajectoryAction = leftParkingTrajectory;
 
         // Get a preview of the trajectory's path:
         Canvas previewCanvas = new Canvas();
@@ -68,30 +61,6 @@ public class ScoringAutoMiddle extends BaseOpMode {
             if (more)
                 MecanumDrive.sendTelemetryPacket(packet);
             telemetry.update();
-        }
-    }
-
-    private void textMenuUpdateUntilComplete(TextMenu textMenu, MenuInput input) {
-        while (!textMenu.isCompleted() && !isStopRequested()) {
-            for (String line : textMenu.toListOfStrings()) {
-                telemetry.addLine(line);
-            }
-            telemetry.update();
-
-            input.update(
-                    gamepad1.left_stick_x, gamepad1.left_stick_y,
-                    gamepad1.dpad_left, gamepad1.dpad_right,
-                    gamepad1.dpad_down, gamepad1.dpad_up,
-                    gamepad1.a
-            );
-            textMenu.updateWithInput(input);
-            sleep(17);
-        }
-    }
-
-    private <E extends Enum> void runIfNotNull (Class<E> enumClass, Runnable runnable) {
-        if (enumClass != null) {
-
         }
     }
 }
