+++ Team6220/src/main/java/org/firstinspires/ftc/team6220/old/ParkAutoRight.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.team6220;
+package org.firstinspires.ftc.team6220.old;
 
 import com.acmerobotics.dashboard.canvas.Canvas;
 import com.acmerobotics.dashboard.telemetry.TelemetryPacket;
@@ -8,8 +8,8 @@ import com.acmerobotics.roadrunner.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
-import org.firstinspires.ftc.team6220.javatextmenu.MenuInput;
-import org.firstinspires.ftc.team6220.javatextmenu.TextMenu;
+import org.firstinspires.ftc.team6220.BaseOpMode;
+import org.firstinspires.ftc.team6220.DRIFTConstants;
 import org.firstinspires.ftc.team6220.roadrunner.MecanumDrive;
 
 /**
@@ -18,27 +18,21 @@ import org.firstinspires.ftc.team6220.roadrunner.MecanumDrive;
  */
 
 @Disabled
-@Autonomous(name="ScoringAutoRight", group="Competition", preselectTeleOp="CompetitionTeleOp")
-public class ScoringAutoRight extends BaseOpMode {
-
+@Autonomous(name="ParkAutoRight", group="Competition", preselectTeleOp="CompetitionTeleOp")
+public class ParkAutoRight extends BaseOpMode {
 
     @Override
     public void runOpMode() {
 
         Pose2d rightPose = DRIFTConstants.RIGHT_STARTING_POSE;
 
-
         MecanumDrive drive = new MecanumDrive(hardwareMap, telemetry, gamepad1, rightPose);
 
-        Action rightScoringTrajectory = drive.actionBuilder(rightPose)
-                .splineTo(new Vector2d(48, 36), (3*Math.PI)/2)
-                .endTrajectory()
-                .splineTo(new Vector2d(48, 50), (5*Math.PI)/4) //scored a sample
-                .splineTo(new Vector2d(25, 10),(3*Math.PI)/2)
-                .splineTo(new Vector2d(28, 10),(1*Math.PI)/1)
+        // Create right parking trajectory
+        Action rightParkingTrajectory = drive.actionBuilder(rightPose)
+                .splineTo(new Vector2d(-60, 60), (Math.PI))
                 .build();
-
-        Action trajectoryAction = rightScoringTrajectory;
+        Action trajectoryAction = rightParkingTrajectory;
 
         // Get a preview of the trajectory's path:
         Canvas previewCanvas = new Canvas();
@@ -68,30 +62,6 @@ public class ScoringAutoRight extends BaseOpMode {
             if (more)
                 MecanumDrive.sendTelemetryPacket(packet);
             telemetry.update();
-        }
-    }
-
-    private void textMenuUpdateUntilComplete(TextMenu textMenu, MenuInput input) {
-        while (!textMenu.isCompleted() && !isStopRequested()) {
-            for (String line : textMenu.toListOfStrings()) {
-                telemetry.addLine(line);
-            }
-            telemetry.update();
-
-            input.update(
-                    gamepad1.left_stick_x, gamepad1.left_stick_y,
-                    gamepad1.dpad_left, gamepad1.dpad_right,
-                    gamepad1.dpad_down, gamepad1.dpad_up,
-                    gamepad1.a
-            );
-            textMenu.updateWithInput(input);
-            sleep(17);
-        }
-    }
-
-    private <E extends Enum> void runIfNotNull (Class<E> enumClass, Runnable runnable) {
-        if (enumClass != null) {
-
         }
     }
 }
