+++ Team6220/src/main/java/org/firstinspires/ftc/team6220/old/ScoringAutoRightPushbot.java
@@ -1,8 +1,10 @@
-package org.firstinspires.ftc.team6220;
+package org.firstinspires.ftc.team6220.old;
 
 import com.acmerobotics.roadrunner.Pose2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+
+import org.firstinspires.ftc.team6220.DRIFTConstants;
 
 @Disabled
 @Autonomous(name = "ScoringAutoRightPushbot", group = "Competition", preselectTeleOp = "CompetitionTeleOp")
